+++ scripts/check_ci_verified_actions.py
@@ -1,0 +1,1 @@
+'ci_verified' not in self.actions['newlabel'] and 'ci_verified' not in self.actions['unlabel']
